Added Following Mechanic; Post liking is still a WIP, currently broken
I added the follower and following mechanic. Rather than go with the
idea of simply adding two columns to the users table with a list of IDs
that the person is following and who is following them, I decided to
use create a new relationships table. This table holds every
relationship and contains a follower_id and then the follow_id. This
makes its much more efficient and practical when it comes to managing
who is following because how much easier it is to looking through the
db. Essentially, a relationship is broken into two categories, passive
and active. A passive is one were you are the one being followed, and
an active one is were you are the one following. To make it possible to
actually follow someone through a button, I made a partial html file
which contains some logic. First it checks if they are looking at their
own profile, to make sure they aren’t going to follow themselves. Then
it checks if you hold an active relationship with the profile you are
on. If you don’t then it shows a button which will create an active
relationship row in the relationship table. You will be listed as the
follower_id and the persons profile you are visiting will be the
following_id. If you already hold an active relationship, the unfollow
button will appear instead. This button will remove the relationship
you hold with the user profile you are visiting.
For up votes, I am using a gem called acts as votable. I have
implemented the database table votes. Not sure exactly how it works but
I ran into a unknown method bug which shouldn’t be happening. Going to
look into how exactly it works tomorrow morning.